feat: apply menu item move to app node, popup a modal when save document

+++ src/main/commands/DeleteKubernetesObjectCommand.ts
@@ -20,6 +20,16 @@ export default class DeleteKubernetesObjectCommand implements ICommand {
     const nodeName: string = node.name;
     const namespace: string = node.getAppNode().namespace;
     const kubeConfig: string = node.getKubeConfigPath();
+    const confirm:
+      | string
+      | undefined = await vscode.window.showInformationMessage(
+      `Delete: ${nodeName}?`,
+      { modal: true },
+      "OK"
+    );
+    if (confirm !== "OK") {
+      return;
+    }
     const result: ServiceResult = await services.deleteKubernetesObject(
       kind,
       nodeName,
